Fix initialize as on new lunar
The logs in .../multiver have a weird line initializing as Player595 at
the start, and then, with a different prefix, initializing as the real
user later.
This new prefix just adds a suffix to one already used, so we now look
for the one that ends at the largest index.