Only run toggle_callback when the user has clicked
We used to run the toggle callback initially when we were setting up the
button. This seemed weird, but was necessary for the button in the
antisniper section, so we make a call to that callback manually now.

This allows us to put the gui code in the antisniper section back into a
more reasonable order.