Fix autodenicking
Autodenicking previously didn't consider the usernames of already
denicked teamates, potentially leading to an already denicked teammate
getting denicked to something else if there was 1 other unknown nick in
the lobby.
Also, we now consider API-denicks to be unknown as well, so that we are
sure that we correctly denick our teammates.