bug: Remove deduplication suffix (e.g. [x2])
Lunar does this, so if a message we want to parse is sent twice in a
row, e.g. if you type /who twice, the last message would be broken, as
we would parse the last username to be 'Player1 [x2]'. We now remove
this prefix so it parses properly again.