feat: Proper bedwars star colors
This rewrites most of the old rating/rendering system to a new one which
allows for multiple colors in one cell, is configurable, and shares more
code between the GUI and terminal outputs.

Also added two settings: rating_configs for configuring how each stat is
rated, and column_order for changing the entries in the stats table and
their order.