fix: Set wants_shown = None on game start and lobby swap
We already set wants_shown = None on joining a queue, but now we also do
it for leaving a queue.
This should not matter in the normal case as wants_shown should either
be False or None in a queue, but if you join late and the overlay gets
confused it can leave the overlay open.