Recommend more stats threads
I'm not sure of the performance implications of too many stats threads
on slow hardware, but due to the GIL we're not going to be hogging too
much cpu anyway. The only potential issue I can think of is if the stats
threads take so much scheduler time that the GUI thread starts lagging.

Increasing the default by 2 should give a decent amount of threads for
any modern hardware (>=4 cores) and where that is too slow it can be
turned down manually.