fix: Don't get stuck in a loop getting stats in rpc
Lunar initializes with Player515 or something similar when you start the
launcher, and then changes to your actual account when you launch
minecraft. Since Player515 doesn't exist, the rpc thread could get stuck
looking up it's stats forever.
To fix this we now check if the username has changed when we fail
getting stats and exit the loop.