Remove AntiSniper denicking
The API uses a now outdated method for denicking, and as such is
returning old and unreliable data. Additionally, denicking like this is
in violation of the Hypixel API Policy.
"Projects should not de-anonymize players such as those who are using
the Hypixel nick feature"

I maintain that the autodenicking still present in the overlay does not
violate this clause as it only works on yourself and your party members,
and only presents information already available to the player.