chore: Add tests for logs from .lunarclient/logs/game
These logs are duplicated from offline/multiver/logs/latest.log, but in
some cases this file fails to write, and the error message appears in
logs/game/<uuid>-master.log

Fixing this is difficult, as we don't want to add all logfiles from the
logs/game directory to the selection menu, as this would be a lot of
options + it would interfere with autoselection when latest.log is also
written to.

The workaround atm is to add the logs/game file manually at each launch.