chore: Use dataReceivedAtMs instead of time.time to compute sessiontime
This means that all data about a player stays static until new data is
requested. I like this, as we now have one consistent view of a player
for the entire duration of the queue. (Unless the queue is super long,
where the stats could expire).

Also use the same calculation when sorting, fixing two bugs!
1. Sort order was reversed due to sorting on last login (session time
   shrinks as last login grows)
2. Sort order used different logic to determine "missing" data