fix: Select correct logfile when switching versions
If the user has used one version a while ago, and then switches to
another version - starting the overlay so that autoselection happens
before the logfile selection menu pops up - the previous logfile would
be selected.

This was due to an incorrect assumption about the list of logfiles being
sorted.