feat: Don't force show the overlay after /who at game start
We want users to be able to do a lobby list when the game starts to
reveal all players. To prevent having to manuall minimize the overlay
after this, we don't set the user's preference to keep it open when done
early in the game (before the overlay is automatically hidden). Lobby
lists at any other time will set the user's preference like usual. This
is important to give users who don't use show on tab a way to manually
show the overlay.