fix: Filter out duplicate players when rendering
If we want to render both Player1 (AmazingNick) and Player1, we don't
render Player1, and just render Player1 (AmazingNick).
This is because duplicate entries in the lobby should not be possible,
except for the case (next commit) where we add all party members to the
lobby. If a party member is nicked, this situation can occur.