chore: Reorganize antisniper settings
- Rename section API key -> AntiSniper
- Move winstreak estimates settings into it
- Move the section down

This used to be a really important setting when we used the
AntiSniper-hosted endpoint as our player stats endpoint. Now that we
have our own dedicated one that doesn't use this key, it is much less
important, and is now more exclusively tied with the winstreaks
estimates setting, as that is currently the only thing it gives you.