chore: Change prompt_if_no_autowho to return True if no autowho
We used to return True if the settings changed, and a flush to disk was
needed. We now return True if the update function gets run, regardless
of whether the settings change or not.
This means we will always flush to disk, but this was already the case,
as `autowho` was missing from `incomplete_settings`.
This also means we can test a more meaningful value which is "did the
user get a prompt", and not "did the settings get updated".