chore: Return player from playerprovider - not playerdata
This moves the parsing of the Hypixel-specific data format into the
player provider, and let's us only care about our domain model `Player`
in the rest of the application code.
This is nice to structure the data earlier, and also simplify switching
to a different API for playerdata (likely a new API on flashlight with a
simpler structure).

Also slightly restructured fetch_get_stats.

NOTE: This move means that we now exclusively use the name returned from
Hypixel on the player, and not the requested username.
This makes sense to me, but we could run into issues where Hypixel has
outdated username data. An option would be to pass in the requested
username as context as before.