chore: Re-enable cache pruning in uv action
This would previously hang, but an issue related to this was fixed in
the v7 release.
This issue cites uv processes running in the background as the reason
for the timeout, which might be the server started in the tls certs test
in our case.

We can re-enable the cache pruning now anyway, but should probably look
into properly cleaning up that server after completing the test.