fix: Use lowercase usernames as cache keys
This fixed an issue discovered while testing that would cause cache
misses if a name had mismatching capitalization from the state and from
the response from the player provider.

I have not seen this happen on Hypixel, but experienced it in testing
with player "Hypixel" in the lobby instead of "hypixel".

This could potentially help prevent issues in the future if Hypixel ever
messes up the capitalization of a name in the chat :^).