feat: Automatically show tags column when relevant
This patch automatically shows the tags column if any player in the
lobby has a tag (any cell value that is not pending, error, nicked, or
empty).

My experience so far is that relatively few players are tagged, such
that showing the column all the time by default would be a waste of
space. Since the tags column is given highest priority when sorting
players, however, it is important that the column is shown when any
players are tagged so that the sorting makes sense.

Users have the option of permanently enabling the tags column if they
prefer that, which also gives them the option of ordering it where they
want it.