fix: Don't poll the Hypixel API
We had a branch in the discord activity updater that would allow a new
request to be made if no activity had occurred during the past 5
minutes. This was implemented to keep the discord activity up to date
even if the overlay fell out of sync with the chatlines or misunderstood
the game end chatline.

An unfortunate consequence of this was that it lead to polling behavior
when users stopped playing but did not close the overlay.
Fixing this here by removing this fallback altogether, as the overlay
should be able to properly pick up game ends.