chore: Move stats queues into controller
This is a core part of the state of the overlay, and should live in the
controller. It also cleans up a lot of function signatures, which is a
good sign.

Idea is to add a behavior for requesting player stats, as this is done a
few different places now.