chore: Delete enqueue_player_request
With get_cached_player_or_enqueue_request we no longer have a need for
just "enqueue request".
There are cases where we want the most up to date stats, but even if we
requested unconditionally we might hit a cache on the backend as well.

Specifically when a game ends, to ensure that the latest stats are
recorded and displayed on discord RPC, we should probably make an
additional request after 1 minute e.g. if we got the stats and it did
not show at least one more game played than the last stats.